Buffer size check before adding to it
It's fairly minor, but if the buffer is already at `buffer_size` then appending to it would be `buffer_size`+1. The newest event would then go out in the next `flush()`.

It really doesn't matter that much, but at least we know that we're always within our own defined limit.