Use the user's celery app
Celery will provide a "default" app if one is not already define.  The
shared_task decorator will use this default app, or the "current" app if
one has already been created by the user.  If the user is already
creating their own app, there is no need to instantiate one specifically
for snowplow.