Do not pin dependencies to specific versions
Because the packages in `setup.py` are really old and they use exact
version, when installing the package into a virtualenv it will end-up
downgrading already existing  packages.

In the Python documentation it is also mentioned that it is not a best
practice to use `install_requires` to pin dependencies to specific versions.

See also:

- https://github.com/snowplow/snowplow-python-tracker/issues/195
- https://github.com/snowplow/snowplow-python-tracker/issues/198
- https://packaging.python.org/discussions/install-requires-vs-requirements/#install-requires